density_estimation/requirements.txt
fastapi
uvicorn
numpy
opencv-python
pillow
requests
ultralytics
shapely
httpx